requirements.txt
Flask==2.2.2
Flask-Cors>=3.0.0
requests==2.28.1
pytest==7.1.2
python-dotenv==0.21.0
